Wave table and PWM variable  type to uint16_t

+++ wave.h
@@ -1,10 +1,11 @@
+#include <stdint.h>
 #define MSTEPS 32
 #define MSTEPS_DIV 5
 #define MSTEPS4 (MSTEPS*4)
-const static int wave_f[]={0,50,100,150,200,249,297,345,392,438,483,526,569,610,650,688,724,
+const static uint16_t wave_f[]={0,50,100,150,200,249,297,345,392,438,483,526,569,610,650,688,724,
                         759,792,822,851,878,903,926,946,964,980,993,1004,1013,1019,1023,1024};
 
-int wave[33];
+uint16_t wave[33];
 void generate_wave(int val);
